backend/requirements.txt
fastapi
uvicorn
openai
pypdf
faiss-cpu
python-multipart
python-dotenv
scikit-learn
